Speed up agent_ui compilation
Extract worktree_service, created_worktrees, worktree_names,
askpass_modal, and the git error toast/output helpers out of git_ui into
a new worktree_service crate. agent_ui and sidebar now depend on it
instead of git_ui, so git_ui and agent_ui compile in parallel.
