+++ crates/agent_ui/src/test_support.rs
@@ -146,7 +146,12 @@ pub async fn record_zed_created_worktree(
 ) {
     let created_at = fake_worktree_created_at(fs, worktree_path).await;
     cx.update(|cx| {
-        git_ui::created_worktrees::record_created_worktree(worktree_path, remote, created_at, cx)
+        worktree_service::created_worktrees::record_created_worktree(
+            worktree_path,
+            remote,
+            created_at,
+            cx,
+        )
     })
     .await
     .unwrap();
