+++ crates/sidebar/src/sidebar.rs
@@ -69,7 +69,7 @@ use workspace::{
     notifications::NotificationId, sidebar_side_context_menu,
 };
 
-use git_ui::worktree_service::{RemoteBranchName, worktree_create_targets};
+use worktree_service::{RemoteBranchName, worktree_create_targets};
 use zed_actions::editor::{MoveDown, MoveUp};
 use zed_actions::{CreateWorktree, NewWorktreeBranchTarget, OpenRecent};
 
@@ -748,7 +748,7 @@ fn create_worktree_in_workspace(
 ) {
     workspace.update(cx, |workspace, cx| {
         let focused_dock = workspace.focused_dock_position(window, cx);
-        git_ui::worktree_service::handle_create_worktree(
+        worktree_service::handle_create_worktree(
             workspace,
             &CreateWorktree {
                 worktree_name: None,
