+++ crates/worktree_service/Cargo.toml
@@ -1,0 +1,56 @@
+[package]
+name = "worktree_service"
+version = "0.1.0"
+edition.workspace = true
+publish.workspace = true
+license = "GPL-3.0-or-later"
+
+[lints]
+workspace = true
+
+[lib]
+name = "worktree_service"
+path = "src/worktree_service.rs"
+
+[dependencies]
+anyhow.workspace = true
+askpass.workspace = true
+collections.workspace = true
+db.workspace = true
+editor.workspace = true
+fs.workspace = true
+futures.workspace = true
+git.workspace = true
+gpui.workspace = true
+language.workspace = true
+log.workspace = true
+menu.workspace = true
+notifications.workspace = true
+project.workspace = true
+proto.workspace = true
+rand.workspace = true
+remote.workspace = true
+remote_connection.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+settings.workspace = true
+terminal.workspace = true
+ui.workspace = true
+util.workspace = true
+workspace.workspace = true
+zed_actions.workspace = true
+zeroize.workspace = true
+
+[dev-dependencies]
+editor = { workspace = true, features = ["test-support"] }
+fs = { workspace = true, features = ["test-support"] }
+git = { workspace = true, features = ["test-support"] }
+gpui = { workspace = true, features = ["test-support"] }
+project = { workspace = true, features = ["test-support"] }
+remote_connection = { workspace = true, features = ["test-support"] }
+settings = { workspace = true, features = ["test-support"] }
+task.workspace = true
+theme.workspace = true
+theme_settings.workspace = true
+workspace = { workspace = true, features = ["test-support"] }
+zlog.workspace = true
