+++ crates/worktree_service/src/created_worktrees.rs
