+++ crates/worktree_service/src/worktree_service.rs
@@ -28,8 +28,12 @@ use git::repository::{FetchOptions, Remote};
 use util::ResultExt as _;
 
 use crate::askpass_modal::AskPassModal;
-use crate::git_panel::{open_output, show_error_toast};
-use crate::worktree_names;
+use crate::notifications::{open_output, show_error_toast};
+
+pub mod askpass_modal;
+pub mod created_worktrees;
+pub mod notifications;
+pub mod worktree_names;
 
 /// A remote-tracking branch reference parsed into its remote and branch parts,
 /// e.g. `origin/main` -> remote `origin`, branch `main`.
